Make overlay+ratio show a pure ratio for 2D histograms.
This has a few disadvantages (mainly that it does not show the actual values at
all) but it makes comparing 2D maps/profiles much easier. Another option would
be to show the ratio "on side", but that needs a new comparison mode. There is
some interference with renderplugins, but with commonly used settings it seems
fine.